Problem: different increments used in strongClone
If you look into `RcSharedPointer`'s `strongClone` and `Weak`'s
`strongClone` you will notice that one uses a regular saturating
increment, while the other one uses a clamped one.

Solution: explain it in the commentary

Bottom line -- it's not a bug (at least I don't think it is :) but an
implication of the environment in which a particular `strongClone`
implementation exists.