Problem: algorithm correctness
Limited coverage of existing tests does not allow us to be sure that the
algorithms used are correct. There may be cases when we're violating
invariants (implicitly or explicitely stated).

Solution: devise a TLA+ model

While it's a first attempt at covering the algorithm, and it's most
definitely an inaccurate approximation (or even a simply buggy one!)
it's nevertheless a great tool to explore its potential transgressions.

As a matter of fact, just trying to come up with the model and running
it through the model checker revealed that the current algorithm is
inadequate, so this commit also fixes a number of issues.

In particular, a virtual weak clone introduced to avoid a race between
deallocation in deinitializing strong clones and deinitializing weak
clones. This idea was shamelessly stolen from Rust's `Arc`.

Also, decrement's previous values are used to test saturation instead of
an error-prone (as model showed) re-reads that try to see if there was
indeed a saturation.

P.S. I don't exclude a probability that the entire model has
significant, debilitating flaws in. As a matter of fact, I am not sure I
got the memory model / ordering baked in. Of particular interest is this
article https://pdfs.semanticscholar.org/fa18/5bd855bd1c894ca176d5f07f9560d6d68d58.pdf
(Verification of Concurrent Data Structures with TLA by Zixian Cai)
to explore this aspect. It would be really interesting to see how far
off is the model from the algorithm implemented in code.