Problem: potential TLA+ specification inconsistency
TLA+ specification assumes that potential deallocation after
decrementing the weak counter in deinitialization is atomically group
with the decrement.

While the topic of memory ordering is never super trivial, atomic
fencing that we have that separates these two operations, suggests that
the specification is likely making an incorrect observation.

Solution: split it into two operations

Luckily, the invariants I was checking for before still (seemingly)
hold.