Problem: use of scratchpads in TLA+ specification
Currently, TLA+ uses a weird "scratchpad" concept that allows some
algorithms to pass some data around steps. It's obscure and it's hard to
follow it.

Solution: drop their usage altogether

DropWeak only used its scratchpad historically and was not utilizing it
now at all. So this one was easy to remove.

DropStrong used its scratchpad, but I've realized that even though it
was using it to store `strong` counter as fetched at that step, it was
also comparing it to 1 to make the program counter change decision, only
for the next step (`deinit`) to check it again. So basically, it was
checking this value twice and effectively there is no reason to pass
this value around. So this allowed to remove strong scratchpad as well.

This simplified the specification and made the model space smaller and I
was able to complete a model check with higher MaxStrongClones and
MaxWeakClones in a reasonable amount of time. Previously, it was taking
far more time as there were more distinct states.