Problem: use of `@fence` in RcSharedPointer
RcSharedPointer is not always expected to be used in multi-threaded
environment. `@fence` might carry overhead that is absolutely not
necessary for a single-threaded program with reference counters.

Solution: abstract it away to `Ops`
and make it an effective no-op in `NonAtomic` ops, as there's always
sequential happens-before ordering in a single thread.