Problem: unclear if we have obvious race conditions
While we do have an approximation of a formal specification (assuming it
is relatively accurate), it's hard to say if the code follows this
specification. So we don't really know if will fail some of the
specification's invariants the moment it hits production.

Solution: stress-test the execution a bit

While in itself an imperfect simulation, this is an attempt to see if
things start falling apart easily or not (again, assuming the tests are
anywhere near being correct!). In this particular instance we test if we
ever deinit or deallocate more than once.

So far it seems to work fine. Take this with a grain of salt. Challenge
and prove it wrong! ;->

Also, some changes were made to the API to accommodate the needs of
tests. However, they were thought to be useful for users of this code.

Firstly, `RcSharedPointer.Weak` type has been exposed publicly to allow
explicit typing of weak clones.

Secondly, `deinit` and `deinitWithCallback` will now return `true` if it
deallocated the value. Previously, they returned nothing.